move exit hooks
